rename: 체크박스 컴포넌트 공용 컴포넌트로 파일 이동
