+++ src/components/home/TaskList/DragabbleCard.tsx
@@ -2,7 +2,7 @@
 import React, { useState } from 'react';
 import { Draggable } from 'react-beautiful-dnd';
 import { Card } from './DragabbleCard.styled';
-import { Checkbox } from './Checkbox';
+import { Checkbox } from '../../common/Checkbox/Checkbox';
 
 function DragabbleCard({ toDo, index }) {
   const [isChecked, setIsChecked] = useState(false);
